+++ runner.py
@@ -3,8 +3,8 @@ Python Script used to read and interpret users input into actions in interpreter
 """
 
 import sys
-from interpreter.interpreter import print_code, execute
-from interpreter.lexer import print_tokens
+from interpreter.min_interpreter import print_code, execute
+from interpreter.min_lexer import print_tokens
 from interpreter.min_parser import print_tree
 
 if __name__ == '__main__':
@@ -15,7 +15,7 @@ if __name__ == '__main__':
         FIRST_ARG = sys.argv[1]
 
         if FIRST_ARG == '--help':
-            print('Usage: python interpreter.py [filename] [flag1] [flag2] ...')
+            print('Usage: python min_interpreter.py [filename] [flag1] [flag2] ...')
             print('Flags available:')
             print('\t-c - show executed code')
             print('\t-l - shot lexer result (raw tokens)')
@@ -27,7 +27,7 @@ if __name__ == '__main__':
             unknown_token = any(flag for flag in flags if flag not in available_flags)
 
             if unknown_token:
-                print('Unknown token: Try typing interpreter.py --help to see usage info')
+                print('Unknown token: Try typing min_interpreter.py --help to see usage info')
             else:
                 if '-c' in flags:
                     print_code(FIRST_ARG)
@@ -40,4 +40,4 @@ if __name__ == '__main__':
                 print("Produced output:")
                 execute(FIRST_ARG)
     except (FileNotFoundError, IndexError):
-        print('Try typing interpreter.py --help to see usage info')
+        print('Try typing min_interpreter.py --help to see usage info')
