+++ tests/test_min_interpreter.py
@@ -1,0 +1,1 @@
+# pylint: skip-file
